DEV: Test site traffic language labels
The Site Traffic language formatter handled browser locale variants without direct unit coverage, leaving script, region, legacy-code, and malformed-input behavior implicit in the system test.\n\nExtract the formatter into an admin utility and cover representative BCP 47 values in one focused unit example. The existing controller remains the only production consumer.
